requirements.txt
# ==============================================================================
# Python Package Requirements for Football Analysis App
# ==============================================================================
# To install all dependencies, run:
# pip install -r requirements.txt
# ==============================================================================

altair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
filelock==3.14.0
filterpy==1.4.5
fonttools==4.51.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lapx==0.5.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pydeck==0.9.0
Pygments==2.18.0
pynndescent==0.5.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
streamlit-drawable-canvas==0.9.3
supervision==0.21.0
sympy==1.12
tenacity==8.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
transformers==4.41.1
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.18
umap-learn==0.5.6
urllib3==2.2.1
watchdog==4.0.0